Add FieldType for categories (#4090)
* adjust typing for datagrid locale

* return nested categories in the same way as pages

* allow selection to be prefetched using expandedIds

* add datagrid type to Selection field type

* fix left side border radius for single cell rows

* remove search field in Selection datagrid type

* configure if items can be delted from Datagrid

* remove deprecation

* remove header from selection field type

* remove obsolete todo comment

* introduce max-height for field type using datagrid

* fix naming in SingleSelection configuration

* expand the category itself only in the datagrid but not in the field type

* add UPGRADE

+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/Datagrid/adapters/ColumnListAdapter.js
@@ -100,21 +100,26 @@ export default class ColumnListAdapter extends AbstractAdapter {
             );
         }
 
-        toolbarItems.push({
-            icon: 'su-cog',
-            type: 'dropdown',
-            options: [
-                {
-                    isDisabled: (index) => {
-                        return activeItems[((index + 1: any): number)] === undefined;
-                    },
-                    label: translate('sulu_admin.delete'),
-                    onClick: (index) => {
-                        onDeleteClick(activeItems[((index + 1: any): number)]);
-                    },
+        const settingOptions = [];
+        if (onDeleteClick) {
+            settingOptions.push({
+                isDisabled: (index) => {
+                    return activeItems[((index + 1: any): number)] === undefined;
                 },
-            ],
-        });
+                label: translate('sulu_admin.delete'),
+                onClick: (index) => {
+                    onDeleteClick(activeItems[((index + 1: any): number)]);
+                },
+            });
+        }
+
+        if (settingOptions.length > 0) {
+            toolbarItems.push({
+                icon: 'su-cog',
+                type: 'dropdown',
+                options: settingOptions,
+            });
+        }
 
         return (
             <div className={columnListAdapterStyles.columnListAdapter}>
